% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{lint_dir_rmd}
\alias{lint_dir_rmd}
\alias{lint_assignment}
\title{Lint all .R and .Rmd recursively in a directory}
\usage{
lint_dir_rmd(...)

lint_assignment(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[lintr:lint_dir]{lintr::lint_dir}}
  \describe{
    \item{\code{path}}{the path to the base directory, by default,
it will be searched in the parent directories of the current directory.}
    \item{\code{relative_path}}{if \code{TRUE}, file paths are printed using their path
relative to the base directory.  If \code{FALSE}, use the full
absolute path.}
    \item{\code{exclusions}}{exclusions for \code{\link[lintr]{exclude}}, relative to the
package path.}
    \item{\code{pattern}}{pattern for files, by default it will take files with .R or .r extension.}
    \item{\code{parse_settings}}{whether to try and parse the settings}
  }}
}
\value{
A list of lint objects.
}
\description{
Apply one or more linters to all of the R files in a directory
}
\section{Functions}{
\itemize{
\item \code{lint_assignment}: exact alias
}}

\examples{
\dontrun{
  lint_dir()
  lint_dir(
    linters = list(semicolon_terminator_linter())
    cache = TRUE,
    exclusions = list("inst/doc/creating_linters.R" = 1, "inst/example/bad.R")
  )
}
}
