% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{lint_package_rmd}
\alias{lint_package_rmd}
\title{Lint all .R and .Rmd package dirs}
\usage{
lint_package_rmd(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[lintr:lint_package]{lintr::lint_package}}
  \describe{
    \item{\code{path}}{the path to the base directory of the package, if \code{NULL},
it will be searched in the parent directories of the current directory.}
    \item{\code{relative_path}}{if \code{TRUE}, file paths are printed using their path
relative to the base directory.  If \code{FALSE}, use the full
absolute path.}
    \item{\code{exclusions}}{exclusions for \code{\link[lintr]{exclude}}, relative to the
package path.}
  }}
}
\value{
A list of lint objects.
}
\description{
Apply one or more linters to all of the R files in a package.
}
\examples{
\dontrun{
  lint_package()

  lint_package(
    linters = with_defaults(semicolon_linter = semicolon_terminator_linter())
    cache = TRUE,
    exclusions = list("inst/doc/creating_linters.R" = 1, "inst/example/bad.R")
  )
}
}
